requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-treebeard==4.7

# Excel processing
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.2.0

# Utils
python-dotenv==1.0.0
tabulate==0.9.0  # Добавлен пакет tabulate

# Development
django-debug-toolbar==4.2.0
black==23.10.1
flake8==6.1.0
isort==5.12.0

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.32.0

# collect
tk>=0.1.0
PySimpleGUI>=4.60.5
tabulate>=0.9.0

python-decouple==3.6
